music_crab: Updates for scheduler 'remoteGlidein'.
When submitting with 'remoteGlidein', there can be more than 500 jobs per task,
but nevertheless you can only submit a maximum of 500 jobs at once.

However, it is not possible to have more than 5000 jobs in one CRAB task
with 'remoteGlidein'. So, the number of events per job is adjusted, so that the
total number of jobs is <= 5000.