music_crab: Do not submit samples with process names that already exist.
music_crab will check the working directory (this is the directory you run
music_crab from, aka. pwd) for CRAB tasks and compare the existing and done ones
with those to be submitted (submit file). If the process name (given in the
submit file) already exists in one of the CRAB tasks, music_crab will give
you a warning and skip the submission.

At this moment the working directory and the 'done' directory in the working directory will be
checked for CRAB tasks.

If you want to resubmit an existing 'process', just move the old one
somewhere else OR use the '--submit' option.