music_crab: Write status of crab jobs to an xml file additionally.
By specifying 'xml_report' in the 'USER' section of the crab config file,
information retrieved by 'crab -status' is written to an xml file in addition to
the regular output on the console.
This xml file might be useful for future reference and enable easier parsing of
job states for our watchdog.