music_crab: Adapted to new 'submit' syntax in crab_2_9_2.
As of crab_2_9_2 you can submit packets of jobs via:
   crab -submit N
where N is the amount of jobs you want to submit. The old syntax, where N is a
range of jobs is also still valid. Yet if you are submitting exactly one job
with number n you must now call:
   crab -submit n,
to distinguish from submitting the first n number of jobs.