Fix: crab_checkwrite used empty stderr of subprocess, now search for specific string.
The crab_checkwrite functions returned false sometimes even though
the test was successful. This behaviour was triggered because the
return value was determined by a check wheter the stderr output from
the subprocess was empty. Depending on the sourced crab3 version it was
possible that crab_checkwrite wrote lines in stderr for sucessful tests.
The new logic searches for a success message string in stdout.