Added aix3adbSampleFinder tool to explore data in aix3adb
I added the tool aix3adbSampleFinder. It can be used to run
simple queries on aix3adb and return the output to the command line.

The main purpose of this script is to serve as a example and
starting point for user created client software for aix3adb

In addition it is possible to see all current fields and their
exact machine name, which may help to write your own scripts.