television.py: Add gridmon support
television now provides a live view of the job stdout, sterr, and others
provided by the gridmon interface. This function can be disabled using an
argument switch, as it is rwth specific. It is only enabled when a
grid certificate passphrase is provided, the job is running at rwth, and
when cmssw is enabled.