watchfrog: add option musicCrab3Input to check 3rd party tasks
The new option musicCrab3Input is added with this commit.
It is now possible to pass a music_crab3 input file and
check if all samples in it are submitted.
It is necessary to also specify the options
--globalTag
--skimmerTag
--jsonFile

In order to have sufficient info for database queries which
can determine if this skimming task has been performed before.